+++ lib/functions/spawn.lua
@@ -1,0 +1,5 @@
+-- Since Lemur doesn't provide an event loop yet, spawn never occurs.
+-- Eventually, we'll have an event loop to implement `spawn` and friends.
+
+return function()
+end