+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Lemur Changelog
 
 ## Master
-* *No changes*
+* Added stub for `spawn` that doesn't do anything
 
 ## v0.1.0 (November 28, 2017)
 * Initial release